separete the container files
