+++ src/components/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import Summary from '../containers/Summary';
-import MoreInfected from '../containers/MoreInfected';
-import MoreNewInfected from '../containers/MoreNewInfected';
+import Summary from '../containers/Summary/Summary';
+import MoreInfected from '../containers/CountriesCases/MoreInfected';
+import MoreNewInfected from '../containers/NewCountriesCases/MoreNewInfected';
 import NavbarApp from './NavbarApp';
 import 'bootstrap/dist/css/bootstrap.min.css';
